requirements.txt
diffusers==0.27.2
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.28.0
safetensors>=0.4.3
Pillow>=10.2.0
python-dotenv>=1.0.1
